requirements.txt
absl-py==2.1.0
accelerate==0.32.1
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.53.0
fsspec==2024.2.0
grpcio==1.64.1
idna==3.7
imageio==2.34.2
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
Markdown==3.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==24.1
pandas==2.2.2
pillow==10.2.0
protobuf==4.25.3
psutil==6.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
six==1.16.0
sympy==1.12
tensorboard==2.17.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.7.2
tokenizers==0.19.*
torch==2.3.*
torch-tb-profiler==0.4.3
torchvision==0.18.*
tqdm==4.66.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
